fixed the conflict after rebased dev close#83
