+++ src/app/data-management/components/batch-submit/batch-submit.component.html
@@ -8,18 +8,20 @@
         </div>
         <div class="card-body">
 
-            <div *ngIf="file">
+            <div class="col" *ngIf="file">
                 <p>文件名:{{ file.name }}</p>
                 <label *ngIf="count  && flag">成功录入{{ count }}条记录！</label>
             </div>
-            <div>
+            <div class="col">
                 <input hidden type="file" #uploadFileInput (change)="onFileSlect($event)" accept=".xlsx"/>
-                <button mat-raised-button type="button" color="warn" *ngIf="!file" (click)="uploadFileInput.click()"
+                <span class="float-right">
+                  <button mat-raised-button class="btn btn-danger" *ngIf="!file" (click)="uploadFileInput.click()"
                   i18n>选择表格</button>
-                <button mat-raised-button type="button" color="warn" *ngIf="file" (click)="uploadFileInput.click()"
+                <button mat-raised-button class="btn btn-danger" *ngIf="file" (click)="uploadFileInput.click()"
                   i18n>重新选择</button>
-                <button mat-raised-button type="button" color="primary" (click)="uploadFile()"
+                <button mat-raised-button class="btn btn-danger" (click)="uploadFile()"
                   [disabled]="!file" i18n>上传表格</button>
+                </span>
             </div>
 
         </div>
