--- src/app/modules/data-management/services/record.service.spec.ts
@@ -1,91 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
-import { Subject } from 'rxjs';
-
-import { RecordService } from './record.service';
-
-import { environment } from 'src/environments/environment';
-import { AUTH_SERVICE } from 'src/app/interfaces/auth-service';
-
-describe('RecordService', () => {
-  let httpTestingController: HttpTestingController;
-  let authenticationSucceed$: Subject<void>;
-
-  beforeEach(() => {
-    authenticationSucceed$ = new Subject<void>();
-    TestBed.configureTestingModule({
-      imports: [
-        HttpClientTestingModule,
-      ],
-      providers: [
-        {
-          provide: AUTH_SERVICE,
-          useValue: {
-            authenticationSucceed: authenticationSucceed$,
-            isAuthenticated: true,
-          },
-        },
-      ]
-    });
-    httpTestingController = TestBed.get(HttpTestingController);
-  });
-
-  afterEach(() => {
-    httpTestingController.verify();
-  });
-
-  it('should be created', () => {
-    const service: RecordService = TestBed.get(RecordService);
-    expect(service).toBeTruthy();
-  });
-
-  it('should delete record', () => {
-    const service: RecordService = TestBed.get(RecordService);
-    const id = 123;
-
-    service.deleteRecord(id).subscribe();
-
-    const req = httpTestingController.expectOne(
-      `${environment.API_URL}/records/${id}/`);
-    expect(req.request.method).toBe('DELETE');
-    req.flush({});
-  });
-
-  it('should get records', () => {
-    const service: RecordService = TestBed.get(RecordService);
-
-    service.getRecords({}).subscribe();
-
-    const url = `${environment.API_URL}/records/?limit=10&offset=0`;
-
-    const req = httpTestingController.expectOne(url);
-    expect(req.request.method).toEqual('GET');
-    req.flush({ count: 2 });
-  });
-
-  it('should get record', () => {
-    const service: RecordService = TestBed.get(RecordService);
-    const id = 1;
-
-    service.getRecord(id).subscribe();
-
-    const url = `${environment.API_URL}/records/${id}/`;
-
-    const req = httpTestingController.expectOne(url);
-    expect(req.request.method).toEqual('GET');
-    req.flush({});
-  });
-
-  it('should batch-submit record', () => {
-    const service: RecordService = TestBed.get(RecordService);
-
-    service.batchSubmitRecord(new File([], 'a.xlsx')).subscribe();
-
-    const url = `${environment.API_URL}/records/actions/batch-submit/`;
-
-    const req = httpTestingController.expectOne(url);
-    expect(req.request.method).toEqual('POST');
-    req.flush({});
-  });
-
-});
