+++ src/app/training-record/training-record.module.ts
@@ -3,7 +3,6 @@ import { NgModule } from '@angular/core';
 import { TrainingRecordRoutingModule } from './training-record-routing.module';
 import { EntryModeComponent } from './components/entry-mode/entry-mode.component';
 import { RecordFormComponent } from './components/record-form/record-form.component';
-import { BatchSubmitComponent } from './components/batch-submit/batch-submit.component';
 import { TrainingRecordComponent } from './training-record.component';
 import { RecordListComponent } from './components/record-list/record-list.component';
 import { RecordDetailComponent } from './components/record-detail/record-detail.component';
@@ -15,7 +14,6 @@ import { SharedModule } from 'src/app/shared/shared.module';
     TrainingRecordComponent,
     EntryModeComponent,
     RecordFormComponent,
-    BatchSubmitComponent,
     RecordListComponent,
     RecordDetailComponent,
     OffCampusEventRecordListComponent,
